==> Modulo_2_2/LIMPIEZA/4_Pair_Prog_Limpieza_4.ipynb <==
# 10/05 - Pair Programming Limpieza 4 - Guadalupe & Lara

Es el momento de ponernos a trabajar con los valores nulos 💪🏽. A lo largo de este ejercicio de pair programming vamos a intentar eliminar los valores nulos de nuestras columnas. En la lección hemos aprendido varios métodos, nosotras os planteamos los ejercicios pero sentiros libres de usar el método que más se adapte a vuestras necesidades. Manos a la obra!


import pandas as pd
import numpy as np

df = pd.read_csv("3.ataques_enproceso3.csv", index_col = 0)

df.head()

- Lo primero que tenemos que evaluar es en que columnas tenemos nulos y que cantidad tenemos en cada una. ¿Hay alguna columna con una gran cantidad de nulos? En caso de que sea así deberemos eliminarla.


nulos = df.isnull().sum()
porcentaje = (nulos/ df.shape[0] * 100).round()

tabla = pd.DataFrame([nulos, porcentaje]).T
tabla.columns = ["nulos", "porcentaje"]
tabla

df["age2"].unique()

**Es el momento de eliminar los nulos:**
- Reemplazad los valores nulos de la columna age por la media de la edad, redondeada a dos decimales.


df["age2"].dtypes

Aunque el ejercicio nos pida que sustituyamos por la media, hemos visto que la media y la mediana tienen una gran diferencia, por lo que haremos los calculos con la mediana.

mediana = df["age2"].median()
mediana

media_edad = df["age2"].median()
media_edad
#media_edad.round(2) #no me funciona .round(2) T_T

df["age2"].unique()

df["age2_fillna"] = df['age2'].fillna(media_edad)

df['age2'].round(2)

df.head() #No puedo verlo u.u

- En relación a la columna de country al tratarse de una columna de tipo categórica, reemplazad los valores nulos por una nueva categória que se llame Unknown.


df["country"].fillna(value = "Unknown", inplace = True)


#Comprobamos cuántos "unknown" hay, que son los 54 nulos sustituidos:
(df["country"] == "Unknown").sum()

- Reemplazad los valores nulos de la columna fatal por Unknown


df["fatal_(y/n)"].isna().sum()

df["fatal_(y/n)"].fillna(value = "Unknown", inplace = True)
#Comprobamos cuántos "unknown" hay, que son los 54 nulos sustituidos:
(df["fatal_(y/n)"] == "Unknown").sum()

df["fatal_(y/n)"].isna().sum()

- Reemplazad los valores nulos de la columna type por el valor más frecuente (la moda)


df["type"].isna().sum()

df["type"].replace(np.nan, df["type"].mode()[0], inplace = True)
#Comprobamos cuántos registros con el valor "unprovocked" hay, que son los que había (4584) más los 8 nans que hemos convertido (4692)
(df["type"] == "Unprovoked").sum()

df["type"].isna().sum()

- Reemplazad los valores nulos de la columna fecha por Unknown


df["date"].replace(np.nan, "Unknown", inplace = True)

- Guardad el csv para seguir trabajando con el en los siguientes ejercicios de pair

df.to_csv("4.Ataques_enproceso_4.csv")
